Made full_board_check Function
Created a function that checks whether the board is full or not.  It uses a for loop to iterate through each position and check if it contains an 'X' or 'O'.  Each time it sees and X or O it adds '1' to the count.  If the count stays under 9 (there are 9 total positions) the function returns False.  If the count reaches 9 the function returns True.